requirements.txt
Flask==3.0.2
Flask-Caching==2.1.0
pandas==2.2.1
numpy==1.26.4
folium==0.19.5
geopy==2.4.1
reportlab==4.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
Flask-WTF==1.2.1
flask-cors==4.0.0

# Testing dependencies
pytest==8.0.2
pytest-cov==4.1.0
coverage==7.4.3
pytest-mock==3.12.0
pytest-flask==1.3.0
